implement LoggerModule
