--- modules/WalkingLogger_module/CMakeLists.txt
@@ -1,5 +1,0 @@
-# Copyright (C) 2019 Fondazione Istituto Italiano di Tecnologia (IIT)
-# All Rights Reserved.
-
-add_subdirectory(LoggerClient)
-add_subdirectory(LoggerServer)
