+++ src/LoggerModule/app/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Copyright (C) 2018 Fondazione Istituto Italiano di Tecnologia (IIT)
+# All Rights Reserved.
+# Authors: Giulio Romualdi <giulio.romualdi@iit.it>
+
+yarp_install(DIRECTORY dcmWalkingLogger DESTINATION ${ICUBCONTRIB_CONTEXTS_INSTALL_DIR})
