+++ src/LoggerModule/app/dcmWalkingLogger/dcmWalkingLogger.ini
@@ -1,0 +1,3 @@
+name               logger
+data_port_name     /data:i
+rpc_port_name      /rpc:i
